+++ routes/api/item-category.php
@@ -1,0 +1,60 @@
+<?php
+
+use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\Route;
+
+Route::group(
+    [
+        'prefix' => Config::get('api.app.version.prefix'),
+        'middleware' => [
+            'convert.route.parameters',
+            'convert.get.parameters'
+        ]
+    ],
+    static function () {
+
+        Route::get(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories',
+            [App\Http\Controllers\View\ItemCategoryController::class, 'index']
+        )->name('item.categories.list');
+
+        Route::options(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories',
+            [App\Http\Controllers\View\ItemCategoryController::class, 'optionsIndex']
+        );
+
+        Route::get(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}',
+            [App\Http\Controllers\View\ItemCategoryController::class, 'show']
+        );
+
+        Route::options(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}',
+            [App\Http\Controllers\View\ItemCategoryController::class, 'optionsShow']
+        );
+
+    }
+);
+
+Route::group(
+    [
+        'prefix' => Config::get('api.app.version.prefix'),
+        'middleware' => [
+            'auth:sanctum',
+            'convert.route.parameters'
+        ]
+    ],
+    static function () {
+
+        Route::post(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories',
+            [App\Http\Controllers\Manage\ItemCategoryController::class, 'create']
+        );
+
+        Route::delete(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}',
+            [App\Http\Controllers\Manage\ItemCategoryController::class, 'delete']
+        );
+
+    }
+);
