+++ routes/api/item-subcategory.php
@@ -1,0 +1,60 @@
+<?php
+
+use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\Route;
+
+Route::group(
+    [
+        'prefix' => Config::get('api.app.version.prefix'),
+        'middleware' => [
+            'convert.route.parameters',
+            'convert.get.parameters'
+        ]
+    ],
+    static function () {
+
+        Route::get(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}/subcategories',
+            [App\Http\Controllers\View\ItemSubcategoryController::class, 'index']
+        );
+
+        Route::options(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}/subcategories',
+            [App\Http\Controllers\View\ItemSubcategoryController::class, 'optionsIndex']
+        );
+
+        Route::get(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}/subcategories/{item_subcategory_id}',
+            [App\Http\Controllers\View\ItemSubcategoryController::class, 'show']
+        );
+
+        Route::options(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}/subcategories/{item_subcategory_id}',
+            [App\Http\Controllers\View\ItemSubcategoryController::class, 'optionsShow']
+        );
+
+    }
+);
+
+Route::group(
+    [
+        'prefix' => Config::get('api.app.version.prefix'),
+        'middleware' => [
+            'auth:sanctum',
+            'convert.route.parameters'
+        ]
+    ],
+    static function () {
+
+        Route::post(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}/subcategories',
+            [App\Http\Controllers\Manage\ItemSubcategoryController::class, 'create']
+        );
+
+        Route::delete(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}/categories/{item_category_id}/subcategories/{item_subcategory_id}',
+            [App\Http\Controllers\Manage\ItemSubcategoryController::class, 'delete']
+        );
+
+    }
+);
