+++ routes/api/item.php
@@ -1,0 +1,65 @@
+<?php
+
+use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\Route;
+
+Route::group(
+    [
+        'prefix' => Config::get('api.app.version.prefix'),
+        'middleware' => [
+            'convert.route.parameters',
+            'convert.get.parameters'
+        ]
+    ],
+    static function () {
+
+        Route::get(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items',
+            [App\Http\Controllers\View\ItemController::class, 'index']
+        )->name('item.list');
+
+        Route::options(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items',
+            [App\Http\Controllers\View\ItemController::class, 'optionsIndex']
+        )->name('item.list.options');
+
+        Route::get(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}',
+            [App\Http\Controllers\View\ItemController::class, 'show']
+        )->name('item.show');
+
+        Route::options(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}',
+            [App\Http\Controllers\View\ItemController::class, 'optionsShow']
+        )->name('item.show.options');
+
+    }
+);
+
+Route::group(
+    [
+        'prefix' => Config::get('api.app.version.prefix'),
+        'middleware' => [
+            'auth:sanctum',
+            'convert.route.parameters'
+        ]
+    ],
+    static function () {
+
+        Route::post(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items',
+            [App\Http\Controllers\Manage\ItemController::class, 'create']
+        );
+
+        Route::delete(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}',
+            [App\Http\Controllers\Manage\ItemController::class, 'delete']
+        );
+
+        Route::patch(
+            'resource-types/{resource_type_id}/resources/{resource_id}/items/{item_id}',
+            [App\Http\Controllers\Manage\ItemController::class, 'update']
+        );
+
+    }
+);
